i2c-new: add partial i2c_xfer implimentation.
Avoiding the use of twi_inter_{on,off} due to very slight optimization
issue.

Concern: the use of pointers as opposed to object and objects as opposed
to pointers in the manner that i2c_transactions are queued up. May run
into excessive copying of data if certain optimizations do not occour.